+++ lib/src/main/kotlin/tkhamez/discordRelay/lib/Gateway.kt
@@ -10,8 +10,11 @@ import io.ktor.client.features.websocket.*
 import io.ktor.client.request.*
 import io.ktor.http.cio.websocket.*
 import kotlinx.coroutines.*
-import java.util.concurrent.BlockingQueue
+import kotlinx.coroutines.channels.Channel
 
+/**
+ * Send when the connection is closed and there will be no attempt to reconnect.
+ */
 const val EVENT_CLOSED = 1
 
 private const val OP_DISPATCH = 0
@@ -22,6 +25,7 @@ private const val OP_INVALID_SESSION = 9
 private const val OP_HELLO = 10
 private const val OP_HEARTBEAT_ACK = 11
 private const val EVENT_READY = "READY"
+private const val EVENT_GUILD_CREATE = "GUILD_CREATE"
 private const val EVENT_RESUMED = "RESUMED"
 private const val EVENT_MESSAGE_CREATE = "MESSAGE_CREATE"
 
@@ -38,18 +42,25 @@ private val httpClient = HttpClient(CIO) {
  * https://discord.com/developers/docs/topics/opcodes-and-status-codes
  */
 class Gateway(
-    private val messages: BlockingQueue<String>,
-    private val events: BlockingQueue<Int>,
-    private val webhook: Webhook
+    private val messages: Channel<String>,
+    private val events: Channel<Int>,
+    private val webhook: Webhook,
 ) {
     private val closeCodesReconnect = setOf<Short>(4000, 4001, 4002, 4003, 4005, 4007, 4008, 4009)
     private val closeCodesStop = setOf<Short>(4004, 4010, 4011, 4012, 4013, 4014)
+
+    private val retryTimeSec: Long = 20
+
+    private var sessionStartLimitRemaining = 1000
+    private var sessionStartCount = 0
 
     private var lastConnectionAttempt: Long = 0
     private var connectingOrConnected = false
 
     private var gatewayResponse: HttpResponseGateway? = null
     private var webSocketSession: DefaultWebSocketSession? = null
+
+    private var connectionJob: Job? = null
 
     private var sequenceNumber: Int? = null
     private var tryResumeConnection = false
@@ -65,49 +76,36 @@ class Gateway(
     private var heartbeatJob: Job? = null
 
     private var retryJob: Job? = null
-    private var closed = false
+    private var isClosed = true
 
-    fun putDummyMessage() {
-        messages.put("")
-    }
+    fun init() {
+        if (connectingOrConnected) {
+            messagesSend(ResString.alreadyConnectingOrConnected)
+            return
+        }
 
-    suspend fun eventsTake(): Int {
-        return withContext(Dispatchers.IO) { events.take() }
-    }
-
-    suspend fun init(): Boolean {
         if (lastConnectionAttempt + 5000 > System.currentTimeMillis()) {
             // It's actually HttpResponseGateway.session_start_limit.max_concurrency times per 5 seconds,
             // which was always 1 for me so far.
-            withContext(Dispatchers.IO) { messages.put(ResString.waitSeconds) }
-            return false
+            messagesSend(ResString.waitSeconds)
+            return
         }
 
-        // Move this up here, it can prevent a "ForegroundServiceDidNotStartInTimeException" on Android if you
-        // click the connect button very fast several times without an internet connection.
-        lastConnectionAttempt = System.currentTimeMillis()
-
-        if (connectingOrConnected) {
-            withContext(Dispatchers.IO) { messages.put(ResString.alreadyConnectingOrConnected) }
-            return false
-        }
         connectingOrConnected = true
 
-        if (!fetchGatewayUrl()) {
-            connectingOrConnected = false
-            // need to send this for Android to stop the service
-            withContext(Dispatchers.IO) { events.put(EVENT_CLOSED) }
-            return false
-        }
-
-        // Connect
-        gatewayResponse?.url?.let { connect(it) }
-
-        return true
+        fullConnect()
     }
 
-    suspend fun close() {
-        close(CloseReason.Codes.NORMAL, CLOSE_REASON_MESSAGE_NORMAL)
+    fun close() {
+        if (retryJob?.isActive == true) {
+            messagesSend(ResString.reconnectAttemptCanceled)
+        }
+        retryJob?.cancel()
+        if (webSocketSession?.isActive == true) {
+            CoroutineScope(Dispatchers.IO).launch { close(CloseReason.Codes.NORMAL, CLOSE_REASON_MESSAGE_NORMAL) }
+        } else {
+            connectingOrConnected = false
+        }
     }
 
     fun testCloseResumeOK() {
@@ -134,11 +132,32 @@ class Gateway(
         CoroutineScope(Dispatchers.IO).launch { send(text) }
     }
 
+    private fun fullConnect() {
+        CoroutineScope(Dispatchers.IO).launch {
+            if (!fetchGatewayUrl()) {
+                connectingOrConnected = false
+                // need to send this for Android to stop the service
+                eventsSendClosed()
+                return@launch
+            }
+
+            // Connect
+            lastConnectionAttempt = System.currentTimeMillis()
+            gatewayResponse?.url?.let { connect(it) }
+        }
+    }
+
     /**
      * Get gateway URL and metadata
      */
     private suspend fun fetchGatewayUrl(): Boolean {
-        withContext(Dispatchers.IO) { messages.put(ResString.requestingGatewayURL) }
+        messagesSend(ResString.requestingGatewayURL)
+
+        if (sessionStartCount + 1 >= sessionStartLimitRemaining) {
+            Config.gatewayResponse = null
+            Config.gatewayResponseBot = mutableMapOf()
+            sessionStartCount = 0
+        }
 
         var gatewayResponseTemp =
             if (Config.isBotToken) Config.gatewayResponseBot[Config.token]
@@ -153,7 +172,7 @@ class Gateway(
                     if (Config.isBotToken) header("Authorization", "Bot " + Config.token)
                 }
             } catch (t: Throwable) {
-                withContext(Dispatchers.IO) { messages.put("${ResString.error}: ${t.message.toString()}") }
+                messagesSend("${ResString.error}: ${t.message.toString()}")
                 return false
             }
             log("HTTP response: $response")
@@ -167,7 +186,7 @@ class Gateway(
         gatewayResponse = gatewayResponseTemp
 
         if (gatewayResponseTemp == null) {
-            withContext(Dispatchers.IO) { messages.put(ResString.failedGatewayURL) }
+            messagesSend(ResString.failedGatewayURL)
             return false
         }
 
@@ -177,11 +196,10 @@ class Gateway(
             Config.gatewayResponse = gatewayResponseTemp
         }
 
-        if ((gatewayResponseTemp.session_start_limit?.remaining ?: 1) < 1) {
-            val resetAfter = gatewayResponseTemp.session_start_limit?.reset_after
-            withContext(Dispatchers.IO) {
-                messages.put(ResString.sessionStartLimitReached.replace("$1", resetAfter.toString()))
-            }
+        sessionStartLimitRemaining = gatewayResponseTemp.session_start_limit?.remaining ?: 1000
+        if (sessionStartLimitRemaining < 1) {
+            val resetAfter = (gatewayResponseTemp.session_start_limit?.reset_after ?: 0) / 1000 / 60
+            messagesSend(ResString.sessionStartLimitReached.replace("$1", resetAfter.toString()))
             return false
         }
 
@@ -189,36 +207,30 @@ class Gateway(
     }
 
     private fun connect(url: String) {
-        connectingOrConnected = true // here also for resume()
+        connectingOrConnected = true // here too for resume()
 
-        if (webSocketSession?.isActive == true) {
-            messages.put(ResString.alreadyConnected)
-            return
-        }
-        messages.put(ResString.connecting)
+        retryJob?.cancel() // for some reason connectionJob does not cancel retryJob
+        heartbeatJob?.cancel()
+        webSocketSession?.cancel()
+        connectionJob?.cancel()
 
-        CoroutineScope(Dispatchers.IO).launch {
+        messagesSend(ResString.connecting)
+
+        connectionJob = CoroutineScope(Dispatchers.IO).launch {
             try {
                 clientClosed = false
                 identifySent = false
-                closed = false
+                isClosed = false
                 httpClient.webSocket(url) {
                     log("WS connected")
                     webSocketSession = this
-                    withContext(Dispatchers.IO) { messages.put(ResString.connected) }
-                    readMessages()
+                    messagesSend(ResString.connected)
+                    readMessages() // this will keep it running
                 }
             } catch (t: Throwable) {
-                val retryTimeSec: Long = 20
-                withContext(Dispatchers.IO) { messages.put(ResString.errorConnection + t.message) }
-                withContext(Dispatchers.IO) {
-                    messages.put(ResString.reconnectAttempt.replace("$1", retryTimeSec.toString()))
-                }
-                trySendResume = false
-                retryJob = CoroutineScope(Dispatchers.IO).launch {
-                    delay(retryTimeSec * 1000)
-                    handleClose()
-                }
+                isClosed = true
+                messagesSend(ResString.errorConnection + t.message)
+                queueHandleClose()
                 return@launch
             }
             webSocketSession?.let {
@@ -238,25 +250,25 @@ class Gateway(
                     try {
                         handleMessage(msg)
                     } catch (t: Throwable) {
-                        withContext(Dispatchers.IO) { messages.put("${ResString.messageError}: ${t.message}") }
+                        messagesSend("${ResString.messageError}: ${t.message}")
                     }
                 }
             }
         } catch (t: Throwable) {
-            withContext(Dispatchers.IO) { messages.put(ResString.errorReceiving + t.message) }
+            messagesSend(ResString.errorReceiving + t.message)
         }
     }
 
-    private suspend fun handleClose(reason: CloseReason? = null) {
-        closed = true
+    private fun handleClose(reason: CloseReason? = null) {
+        isClosed = true
+        retryJob?.cancel()
+
         val closeCode: Short = reason?.code ?: 0
         val closeMessage: String = reason?.message ?: ""
 
         if (closeCode > 0) { // not when trying to reconnect after a delay
             val who = if (clientClosed) ResString.client else ResString.server
-            withContext(Dispatchers.IO) {
-                messages.put("${ResString.disconnected} ($who, $closeCode $closeMessage).")
-            }
+            messagesSend("${ResString.disconnected} ($who, $closeCode $closeMessage).")
         }
 
         if (
@@ -267,7 +279,7 @@ class Gateway(
         } else if (!tryResumeConnection) {
             sessionId = null
             connectingOrConnected = false
-            withContext(Dispatchers.IO) { events.put(EVENT_CLOSED) }
+            eventsSendClosed()
         }
         tryResumeConnection = false
     }
@@ -275,23 +287,42 @@ class Gateway(
     private suspend fun close(code: CloseReason.Codes, reason: String = "") {
         clientClosed = true
         if (!tryResumeConnection && retryJob?.isActive == true) {
-            withContext(Dispatchers.IO) { messages.put(ResString.reconnectAttemptCanceled) }
+            messagesSend(ResString.reconnectAttemptCanceled)
         }
+
         retryJob?.cancel()
+        if (tryResumeConnection) {
+            queueHandleClose() // handleClose() will not be called without Internet connection
+        }
+
         webSocketSession?.close(CloseReason(code, reason))
+
         connectingOrConnected = false
-        // It's not necessary to send EVENT_CLOSED here.
         log("WS closed ($code)")
     }
 
-    private suspend fun resumeTheConnection() {
+    private fun queueHandleClose() {
+        messagesSend(ResString.reconnectAttempt.replace("$1", retryTimeSec.toString()))
+        trySendResume = false
+        retryJob?.cancel()
+        retryJob = CoroutineScope(Dispatchers.IO).launch {
+            delay(retryTimeSec * 1000)
+            handleClose()
+        }
+    }
+
+    private fun resumeTheConnection() {
         trySendResume = true
         gatewayResponse?.url?.let {
-            connect(it)
+            if (sessionStartCount + 1 >= sessionStartLimitRemaining) {
+                fullConnect()
+            } else {
+                connect(it)
+            }
         } ?: run {
             connectingOrConnected = false
-            withContext(Dispatchers.IO) { messages.put(ResString.cannotResume) }
-            withContext(Dispatchers.IO) { events.put(EVENT_CLOSED) }
+            messagesSend(ResString.cannotResume)
+            eventsSendClosed()
         }
     }
 
@@ -312,11 +343,11 @@ class Gateway(
     @Throws(JsonSyntaxException::class)
     private suspend fun handleMessage(messageText: String) {
         val gson = Gson()
-        val message = gson.fromJson(messageText, WsReceive::class.java)
+        val message = gson.fromJson(messageText, WebSocketReceiveMessage::class.java)
         if (message.op == OP_HELLO) {
-            val messageHello = gson.fromJson(gson.toJson(message.d), WsReceiveHello::class.java)
+            val messageHello = gson.fromJson(gson.toJson(message.d), WebSocketReceiveHello::class.java)
             if (messageHello.heartbeat_interval == null) {
-                withContext(Dispatchers.IO) { messages.put(ResString.invalidHello) }
+                messagesSend(ResString.invalidHello)
                 return
             }
             startHeartbeats(messageHello.heartbeat_interval)
@@ -340,19 +371,23 @@ class Gateway(
         } else if (message.op == OP_DISPATCH) {
             sequenceNumber = message.s
             if (message.t == EVENT_READY) {
-                val messageReady = gson.fromJson(gson.toJson(message.d), WsReceiveReady::class.java)
+                val messageReady = gson.fromJson(gson.toJson(message.d), WebSocketReceiveReady::class.java)
                 if (messageReady.session_id != null) {
                     sessionId = messageReady.session_id
-                    withContext(Dispatchers.IO) { messages.put(ResString.ready) }
+                    messagesSend(ResString.ready)
+                    storeChannelInfo(messageReady.guilds)
                 } else {
-                    withContext(Dispatchers.IO) { messages.put(ResString.readyError) }
+                    messagesSend(ResString.readyError)
                     tryResumeConnection = true
                     close(CloseReason.Codes.PROTOCOL_ERROR, CLOSE_REASON_MESSAGE_PROTOCOL_ERROR)
                 }
+            } else if (message.t == EVENT_GUILD_CREATE) {
+                val messageGuildCreate = gson.fromJson(gson.toJson(message.d), DiscordGuild::class.java)
+                storeChannelInfo(listOf(messageGuildCreate))
             } else if (message.t == EVENT_RESUMED) {
-                withContext(Dispatchers.IO) { messages.put(ResString.connectionResumed) }
+                messagesSend(ResString.connectionResumed)
             } else if (message.t == EVENT_MESSAGE_CREATE) {
-                val messageMessageCreate = gson.fromJson(gson.toJson(message.d), WsReceiveMessageCreate::class.java)
+                val messageMessageCreate = gson.fromJson(gson.toJson(message.d), ChannelMessage::class.java)
                 relayMessage(messageMessageCreate)
             }
         }
@@ -363,18 +398,24 @@ class Gateway(
         heartbeatLastAck = 0
         var heartbeatLastSend: Long = 0
         heartbeatJob = CoroutineScope(Dispatchers.IO).launch {
-            while (webSocketSession?.isActive == true) {
+            while (isActive && webSocketSession?.isActive == true) {
                 // Wait first
                 var sleep = heartbeatInterval
-                while (webSocketSession?.isActive == true && sleep > 0) {
-                    delay(200.toLong().coerceAtMost(sleep))
-                    sleep -= 200
+                val waitInterval = 200L
+                while (isActive && webSocketSession?.isActive == true && sleep > 0) {
+                    delay(waitInterval.coerceAtMost(sleep))
+                    sleep -= waitInterval
+                }
+
+                if (webSocketSession?.isActive != true) {
+                    break
                 }
 
                 // Check last acknowledge
                 if (heartbeatLastSend > heartbeatLastAck) {
                     // probably failed or "zombied" connection, terminate and resume
                     tryResumeConnection = true
+                    messagesSend("${ResString.closing} (${CloseReason.Codes.PROTOCOL_ERROR})")
                     close(CloseReason.Codes.PROTOCOL_ERROR, CLOSE_REASON_MESSAGE_PROTOCOL_ERROR)
                     break
                 }
@@ -384,10 +425,10 @@ class Gateway(
                 heartbeatLastSend = System.currentTimeMillis()
             }
 
-            if (!clientClosed && !closed && retryJob?.isActive != true) {
+            if (!clientClosed && !isClosed && retryJob?.isActive != true) {
                 // e.g. when the internet connection was lost the socket is not active but the close
                 // event did not fire. Calling close() now doesn't work.
-                withContext(Dispatchers.IO) { messages.put(ResString.connectionLost) }
+                messagesSend(ResString.connectionLost)
                 resumeTheConnection()
             }
         }
@@ -395,22 +436,23 @@ class Gateway(
 
     private suspend fun sendHeartbeat() {
         val gson = GsonBuilder().serializeNulls().create()
-        send(gson.toJson(WsSendHeartbeat(op = OP_HEARTBEAT, d = sequenceNumber)))
+        send(gson.toJson(WebSocketSendHeartbeat(op = OP_HEARTBEAT, d = sequenceNumber)))
     }
 
     private suspend fun sendIdentify() {
+        sessionStartCount ++
         val wait = identifyLastSend + 5000 - System.currentTimeMillis()
         if (wait > 0) {
             log("'Identify' too fast, wait: $wait ms")
             delay(wait)
         }
         identifyLastSend = System.currentTimeMillis()
-        val identify = WsSendIdentify(
+        val identify = WebSocketSendIdentify(
             op = OP_IDENTIFY,
-            d = WsSendIdentifyData(
+            d = WebSocketSendIdentifyData(
                 token = Config.token,
-                intents = 512, // GUILD_MESSAGES - MESSAGE_CREATE, MESSAGE_UPDATE, MESSAGE_DELETE, MESSAGE_DELETE_BULK
-                properties = WsSendIdentifyDataProperties("Unknown", "discord-relay", "discord-relay"),
+                intents = 513, // GUILDS (1) and GUILD_MESSAGES (512)
+                properties = WebSocketSendIdentifyDataProperties("Unknown", "discord-relay", "discord-relay"),
             ),
         )
         send(Gson().toJson(identify))
@@ -421,9 +463,9 @@ class Gateway(
             sendIdentify()
             return
         }
-        val resume = WsSendResume(
+        val resume = WebSocketSendResume(
             op = OP_RESUME,
-            d = WsSendResumeData(
+            d = WebSocketSendResumeData(
                 token = Config.token,
                 session_id = sessionId ?: "",
                 seq = sequenceNumber,
@@ -432,15 +474,40 @@ class Gateway(
         send(Gson().toJson(resume))
     }
 
-    private suspend fun relayMessage(messageMessageCreate: WsReceiveMessageCreate) {
+    private fun storeChannelInfo(guilds: List<DiscordGuild>?) {
+        if (guilds == null) {
+            return
+        }
+        for (guild in guilds) {
+            if (guild.unavailable == true || guild.channels == null) {
+                continue
+            }
+            val tmpGuild = DiscordGuild(id = guild.id, name = guild.name)
+            for (channel in guild.channels) {
+                if (Config.channelIdList().contains(channel.id)) {
+                    Config.channels.add(GuildChannel(id = channel.id, name = channel.name, guild = tmpGuild))
+                }
+            }
+        }
+    }
+
+    private suspend fun relayMessage(message: ChannelMessage) {
         if (
-            !Config.channelIdList().contains(messageMessageCreate.channel_id) ||
-            (Config.onlyMentionEveryone && messageMessageCreate.mention_everyone != true)
+            !Config.channelIdList().contains(message.channel_id) ||
+            (Config.onlyMentionEveryone && message.mention_everyone != true)
         ) {
             return
         }
 
-        withContext(Dispatchers.IO) { messages.put(ResString.messageReceived) }
-        webhook.sendMessage(messageMessageCreate)
+        messagesSend(ResString.messageReceived)
+        webhook.sendMessage(message)
+    }
+
+    private fun messagesSend(message: String) {
+        CoroutineScope(Dispatchers.Default).launch { messages.send(message) }
+    }
+
+    private fun eventsSendClosed() {
+        CoroutineScope(Dispatchers.Default).launch { events.send(EVENT_CLOSED) }
     }
 }