+++ lib/src/main/kotlin/tkhamez/discordRelay/lib/functions.kt
@@ -1,28 +1,29 @@
 package tkhamez.discordRelay.lib
 
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
+import kotlinx.coroutines.channels.Channel
 import java.text.SimpleDateFormat
 import java.util.*
-import java.util.concurrent.BlockingQueue
-import java.util.concurrent.LinkedBlockingQueue
 
-private var messages: BlockingQueue<String> = LinkedBlockingQueue()
+private var messages = Channel<String>(5000)
+private var events = Channel<Int>(100)
 
-private var gateway = Gateway(messages, LinkedBlockingQueue(), Webhook(messages))
+private var gateway = Gateway(messages, events, Webhook(messages))
 
-@Suppress("SimpleDateFormat")
-fun log(message: Any?) {
-    val now = SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS").format(Date())
-    println("$now $message")
+suspend fun messagesReceive(): String {
+    return messages.receive()
 }
 
-suspend fun messagesTake(): String {
-    return withContext(Dispatchers.IO) {
-        messages.take()
-    }
+suspend fun eventsReceive(): Int {
+    return events.receive()
 }
 
 fun getGateway(): Gateway {
     return gateway
+}
+
+@Suppress("SimpleDateFormat")
+fun log(message: Any?) {
+    val now = SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS").format(Date())
+    val msg = message.toString().replace(Config.token, "*****")
+    println("$now $msg")
 }